+++ inc/ray.h
@@ -6,7 +6,7 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 20:15:36 by wlin              #+#    #+#             */
-/*   Updated: 2024/12/24 17:30:52 by wlin             ###   ########.fr       */
+/*   Updated: 2024/12/25 23:53:06 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ t_view	init_view_params(t_cam camera);
 void	my_mlx_pixel_put(t_img *img, int x, int y, int color);
 void	render_scene(t_data *data);
 t_ray	generate_ray(t_cam camera, t_view view_params, int x, int y);
-int		normal_color(double normal[3]);
+int		normal_color(float normal[3]);
 // int		ray_color(t_ray ray, t_obj obj, double t);
 
 #endif
